refactor(LottoGenerator): Impl 컨벤션 대신 의미있는 네이밍 사용
