+++ src/test/java/lotto/domain/LottoGeneratorTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static lotto.domain.LottoGeneratorImpl.LOTTO_NUMBER_COUNT;
+import static lotto.domain.RandomLottoGenerator.LOTTO_NUMBER_COUNT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
@@ -14,7 +14,7 @@ public class LottoGeneratorTest {
     @Test
     @DisplayName("1 ~ 45 숫자 6개로 구성된 숫자세트 생성 작업이 정상적으로 동작한다.")
     public void constructor() {
-        LottoGeneratorImpl lottoGenerator = new LottoGeneratorImpl();
+        RandomLottoGenerator lottoGenerator = new RandomLottoGenerator();
         List<Integer> lottoNumbers = lottoGenerator.create();
         assertThat(lottoNumbers).hasSize(LOTTO_NUMBER_COUNT);
         assertAll(
