+++ src/test/java/lotto/domain/LottoTest.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
 
-import static lotto.domain.LottoGeneratorImpl.LOTTO_NUMBER_COUNT;
+import static lotto.domain.RandomLottoGenerator.LOTTO_NUMBER_COUNT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
